Bug 1609266 - Support group reviewers r?dkl
The existing implementation assumed that each reviewer was an individual user. However, it's possible to assign groups ("projects") as reviewers.
In this case, each user in the group should be emailed.

Note that it's possible for a user to be in more than one reviewing group, and be a recipient multiple times for the same event.
This issue is captured by bug 1632919.